Support for several feature flags
