+++ modules/llrt_os/Cargo.toml
@@ -7,6 +7,12 @@ license = "Apache-2.0"
 repository = "https://github.com/awslabs/llrt"
 readme = "README.md"
 
+[features]
+default = []
+
+network = ["sysinfo/network"]
+statistics = []
+
 [dependencies]
 home = "0.5"
 llrt_utils = { version = "0.3.0-beta", path = "../../libs/llrt_utils", default-features = false }
@@ -15,9 +21,8 @@ once_cell = "1"
 parking_lot = "0.12"
 rquickjs = { git = "https://github.com/DelSkayn/rquickjs.git", rev = "3af3f46b13eb89a2694e5e4e2e73924a20fa9dd1", default-features = false }
 sysinfo = { version = "0.32", features = [
-  "network",
   "system",
-], default-features = false }
+],default-features = false }
 
 [target.'cfg(unix)'.dependencies]
 libc = "0.2"
